Add Details page save functionality
